move tests to canonical directory
